validator: split the proxy CDC group into the proxy module
cdc declares two groups, one reading the database directly and one
through scylla-proxy, so the file spanned two clusters and could not sit
on either side of the boundary.

Split the proxy group, its fixture and its one test out into proxy::cdc;
no change to either.
