+++ crates/validator/src/proxy/mod.rs
@@ -3,5 +3,6 @@
  * SPDX-License-Identifier: LicenseRef-ScyllaDB-Source-Available-1.1
  */
 
+mod cdc;
 mod full_scan;
 mod routing;
